Refactor to move generate step into separate file
